problem: excluded tests contain required code for non-excluded tests
solution: refactor utils and test data to adjacent files
